// src/app/models/historia.ts
export interface Historia {
    nombre: string;
    folio: number;
    domicilio: string;
    fecha: string;
    edad: number;
    ocupacion: string;
    telpar: string;
    telofi: string;
    dirofi: string;
    recomen: string;
    h1: string;
    h2: string;
    h2a: string;
    h3: string;
    h4: string;
    h4a: string;
    h5: string;
    h6: string;
    h7: string;
    h7a: boolean;
    h7b: boolean;
    h7c: boolean;
    h7d: boolean;
    h7e: boolean;
    h7f: boolean;
    h7g: boolean;
    h7h: boolean;
    h7i: boolean;
    h7j: boolean;
    h8: string;
    h8a: string;
    h9: string;
    h10: string;
    firma: string;
    da1: string;
    da2: string;
    da3: string;
    da4: string;
    da5: string;
    da6: string;
    da7: string;
    da8: string;
    da9: string;
    de1: string;
    de2: string;
    de3: string;
    de4: string;
    de5: string;
    di1: string;
    di2: string;
    di3: string;
    di4: string;
    di5: string;
    di6: string;
    di7: string;
    di8: string;
    di9: string;
    di10: string;
    dp1: string;
    dp2: string;
    dp3: string;
    dp4: string;
    dp5: string;
    dp6: string;
    dr1: string;
    dr2: string;
    dr3: string;
    dr4: string;
    dr5: string;
    dr6: string;
    dr7: string;
    dr8: string;
    dr9: string;
    dr10: string;
    dr11: string;
    dr12: string;
    dr13: string;
    dr14: string;
    dr15: string;
    dr16: string;
    dr17: string;
    dd1: string;
    dd2: string;
    dd3: string;
    dd4: string;
    dd5: string;
    dd6: string;
    dd7: string;
    dd8: string;
    dt1: string;
    dt2: string;
    dt3: string;
    dt4: string;
    dt5: string;
    dt6: string;
    dt7: string;
    dt8: string;
    dt9: string;
    dt10: string;
    dt11: string;
    dpn1: string;
    dpn2: string;
    dpn3: string;
    dpn4: string;
    diente: string;
    fecha2: string;
    diente2: string;
    grapa: string;
    fechap1: string;
    fechap2: string;
    fechap3: string;
    fechap4: string;
    fechap5: string;
    descp1: string;
    descp2: string;
    descp3: string;
    descp4: string;
    descp5: string;
    observa: string;
    mm1: string;
    mm2: string;
    mm3: string;
    mm4: string;
    mm5: string;
    mm6: string;
    mm7: string;
    mm8: string;
    mm9: string;
    re1: string;
    re2: string;
    re3: string;
    re4: string;
    re5: string;
    re6: string;
    re7: string;
    re8: string;
    re9: string;
    la1: string;
    la2: string;
    la3: string;
    la4: string;
    la5: string;
    la6: string;
    la7: string;
    la8: string;
    la9: string;
    te1: string;
    te2: string;
    te3: string;
    te4: string;
    te5: string;
    te6: string;
    te7: string;
    te8: string;
    te9: string;
    lm1: string;
    lm2: string;
    lm3: string;
    lm4: string;
    lm5: string;
    lm6: string;
    lm7: string;
    lm8: string;
    lm9: string;
    img1: string;
    img2: string;
    img3: string;
    img4: string;
    desc1: string;
    desc2: string;
    desc3: string;
    desc4: string;
}