// src/app/models/expediente.ts
export interface Expediente {
    nombre: string;
    folio: number;
    edad: number;
    tel: string;
    fecha: string;
    sexo: string;
    cel: string;
    email: string;
    direcc: string;
    cp: string;
    colonia: string;
    munic: string;
    estado: string;
    escola: string;
    ecivil: string;
    fechan: string;
    peso: string;
    estatu: string;
    g1: string;
    g2: string;
    g3: string;
    g4: string;
    g5: string;
    g6: string;
    g7: string;
    g1a: string;
    g2a: string;
    g3a: string;
    g4a: string;
    g5a: string;
    g6a: string;
    g7a: string;
    e1: string;
    e2: string;
    e3: string;
    e4: string;
    e5: string;
    e6: string;
    e7: string;
    e8: string;
    e9: string;
    e10: string;
    ea1: string;
    ea2: string;
    ea3: string;
    ea4: string;
    ea5: string;
    ea6: string;
    ea7: string;
    ea8: string;
    ea9: string;
    ea10: string;
    h1: string;
    h2: string;
    h3: string;
    h4: string;
    h5: string;
    h1a: string;
    h2a: string;
    h3a: string;
    h4a: string;
    h5a: string;
    fultima: string;
    trataa: string;
    aditame: string;
    vcepill: string;
    od1: string;
    od2: string;
    od3: string;
    od4: string;
    od5: string;
    od6: string;
    od7: string;
    od8: string;
    od9: string;
    od10: string;
    od11: string;
    od12: string;
    od13: string;
    od14: string;
    od15: string;
    od16: string;
    od17: string;
    od18: string;
    od19: string;
    od20: string;
    od21: string;
    od22: string;
    od23: string;
    od24: string;
    od25: string;
    od26: string;
    od27: string;
    od28: string;
    od29: string;
    od30: string;
    od31: string;
    od32: string;
    od33: string;
    od34: string;
    od35: string;
    od36: string;
    od37: string;
    od38: string;
    od39: string;
    od40: string;
    od41: string;
    od42: string;
    od43: string;
    od44: string;
    od45: string;
    od46: string;
    od47: string;
    od48: string;
    od49: string;
    od50: string;
    od51: string;
    od52: string;
    f1: string;
    f2: string;
    f3: string;
    f4: string;
    f5: string;
    f6: string;
    f7: string;
    f8: string;
    f9: string;
    f10: string;
    f11: string;
    dt1: string;
    dt2: string;
    dt3: string;
    dt4: string;
    dt5: string;
    dt6: string;
    dt7: string;
    dt8: string;
    dt9: string;
    dt10: string;
    dt11: string;
    dg1: string;
    dg2: string;
    dg3: string;
    dg4: string;
    dg5: string;
    dg6: string;
    dg7: string;
    dg8: string;
    dg9: string;
    dg10: string;
    dg11: string;
    t1: string;
    t2: string;
    t3: string;
    t4: string;
    t5: string;
    t6: string;
    t7: string;
    t8: string;
    t9: string;
    t10: string;
    t11: string;
    fi1: string;
    fi2: string;
    fi3: string;
    fi4: string;
    fi5: string;
    fi6: string;
    fi7: string;
    fi8: string;
    fi9: string;
    fi10: string;
    fi11: string;
    obser: string;
    firma1: string;
    firma2: string;
    img1: string;
    img2: string;
    img3: string;
    img4: string;
    img5: string;
    img6: string;
    img7: string;
    img8: string;
    des1: string;
    des2: string;
    des3: string;
    des4: string;
    des5: string;
    des6: string;
    des7: string;
    des8: string;
}